solve problem 13 part 2
I gave up on this one and read examples of how people solved it, which
involved learning about [Chinese remainder theorem][0], which I didn't
really understand ... but the sieve method did make some sense and was
easy to implement mostly going off of intuition.

I feel like I was close to the sieve method in the scratchpad of notes
that I was making earlier in the code, but the roadblack I was hitting
was trying to find a way to compute the first `t` value exactly instead
of just starting at 0 and counting forward by `buses[0]` until hitting
the start of the "pattern" for the first 2 buses before increasing the
sieve (I was calling this the "cycle" value in my notes).

[0]: https://en.wikipedia.org/wiki/Chinese_remainder_theorem