problem15 part 2
It takes 17-18 seconds to run `solve` on my input for 30 million rounds.

The initial approach of only storing what is absolutely necessary seems
like it came in handy - I'm not sure if there is a more optimal
algorithm, but solving part 2 in less than 20 seconds is good enough for
me.