16-2: remove unused code, fix answer
I was reading it wrong:

```
Once you work out which field is which, look for the six fields on
your ticket that start with the word departure. What do you get if you
multiply those six values together?
```

I read this originally as return the product of the positions in the
ticket that the fields that start with departure appear, but it was
asking for the product of *my ticket's* values for those fields.