problem18: parts 1 and 2 done
This won't wasn't that much fun. In reading it, it reminded me of
"reverse polish notation" from college, and in googling the phrase to
double-check if I was remembering the name of the concept correctly, I
more or less spoiled how to solve it by re-reading about
https://en.wikipedia.org/wiki/Shunting-yard_algorithm.

It didn't feel like it would be any fun to re-discover that from first
principles, so I basically just implemented the algorithm from the
Wikipedia page.