problem 19 part 2 done
I only solved it in the special case of the new rules, and not
generically.

Rewrote the new rules to avoid loops:

```
8: 42 | 42 8
11: 42 31 | 42 11 31
```

Rule 8 can be understood as "match rule 42 one or more times", and "one
or more times" in regex terms is a "+" character.

Rule 11 is a little tougher - it seems to say repeat rule 42 one or more
times, and then repeat rule 31 the same number of times. I don't think
there is a way to express "the same number of times" in a regex, so I
did a hack: rewrite this to be:

```
11: 42 31 | 42 42 31 31 | 42 42 42 31 31 31 | ...
```

up to 20 times. It works!