solved problem 23 part 2
rewrote the algorithm using a list to store all the cups in (with logic
to wraparound when needed) to instead use a dict that stored the next
cup for each cup, so if cup 1 is immediately before cup 3, `d[1] == 3`.

This is essentially a linked list, minus the need/support for retrieving
the item at a given index, which is not even defined for a circular list
of cups like in the problem.

Takes the runtime down to around 20-30 seconds, seems like it should be
faster though - would be interesting to optimize better.