2021 day 15: rewrite as priority queue with heapq
I feel bad about this one, I got halfway there with realizing I need
refactor the code to use a priority queue / heap to figure out what node
to visit next, instead of examining every node in the unvisited set and
finding the minimum dist[n] - but I had trouble wrapping my head around
how to _modify_ the priority of items already in the heap (the cost of
each neighbor after adjusting it).

I ended up finding a few solutions online to this problem and basing my
modifications off of that.

I think the reason that this works without actually modifying items once
they are in the queue is that we add a neighbor potentially multiple
times, and then ignore any items popped off the queue by checking if
they are already in the visited set.